refactor(web): move filter selects into filters folder
